requirements.txt
Django==4.2.16
djangorestframework==3.15.2
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
matplotlib==3.9.0
weasyprint==62.3
plotly==5.22.0
python-dateutil==2.9.0.post0
django-environ==0.11.2
reportlab==4.2.5

